update to twilight rc-0.16

+++ assyst-core/src/assyst.rs
@@ -45,7 +45,7 @@ pub struct Assyst {
 impl Assyst {
     pub async fn new() -> anyhow::Result<Assyst> {
         let http_client = Arc::new(HttpClient::new(CONFIG.authentication.discord_token.clone()));
-        let shard_count = http_client.gateway().authed().await?.model().await?.shards;
+        let shard_count = http_client.gateway().authed().await?.model().await?.shards as u64;
         let database_handler = Arc::new(RwLock::new(
             DatabaseHandler::new(CONFIG.database.to_url(), CONFIG.database.to_url_safe()).await?,
         ));
